+++ app/src/androidTest/java/com/example/remembrall/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.remebrall
+package com.example.remembrall
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
