+++ app/src/main/java/com/example/remembrall/read/ReadDiaryFragment.kt
@@ -1,11 +1,11 @@
-package com.example.remebrall.read
+package com.example.remembrall.read
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.example.remebrall.R
+import com.example.remembrall.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
