+++ app/src/test/java/com/example/remembrall/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.remebrall
+package com.example.remembrall
 
 import org.junit.Test
